+++ compare-frontend/src/app/modules/wizard/converter.service.ts
@@ -28,3 +28,37 @@ export interface Converter {
 }
 
 export const converterToken = new InjectionToken<Converter>('WizardConverter');
+
+
+export class DefaultConverter implements Converter {
+    private step: string = "";
+
+    applies(scenario: ScenarioName, step: string): boolean {
+        // Sneaky hack: setting step so when a call to `getInfo` comes next, we know what data to fetch.
+        this.step = step;
+        return true;
+    }
+
+    getInfo(state: RiesgosScenarioState, data: RiesgosProductResolved[]): WizardComposite {
+        const step = state.steps.find(s => s.step.id === this.step)!;
+        const inputs = step.step.inputs;
+
+        const inputOptions: WizardComposite["inputs"] = [];
+        for (const input of inputs) {
+            const inputValue = data.find(d => d.id === input.id);
+            inputOptions.push({
+                label: input.id,
+                formtype: input.options ? 'string-select' : 'string',
+                currentValue: inputValue,
+                options: input.options
+            });
+        }
+
+        return {
+            hasFocus: false,
+            step: step,
+            inputs: inputOptions,
+        }
+    }
+
+}